deleteMessage functionality done

+++ src/main/java/DAO/MessageDAO.java
@@ -77,27 +77,34 @@ public class MessageDAO {
 
     public Message deleteMessageById(int message_id) {
         Connection connection = ConnectionUtil.getConnection();
-
+    
         try {
-            String sql = "DELETE FROM Message WHERE message_id = ?";
-            PreparedStatement ps = connection.prepareStatement(sql);
-            ps.setInt(1, message_id);
-
-            ResultSet rs = ps.executeQuery();
-
-            while(rs.next()){
+            String selectSql = "SELECT * FROM Message WHERE message_id = ?";
+            PreparedStatement selectPs = connection.prepareStatement(selectSql);
+            selectPs.setInt(1, message_id);
+    
+            ResultSet rs = selectPs.executeQuery();
+    
+            if (rs.next()) {
                 int posted_by = rs.getInt("posted_by");
                 String message_text = rs.getString("message_text");
                 Long time_posted_epoch = rs.getLong("time_posted_epoch");
+    
+                String deleteSql = "DELETE FROM Message WHERE message_id = ?";
+                PreparedStatement deletePs = connection.prepareStatement(deleteSql);
+                deletePs.setInt(1, message_id);
+                deletePs.executeUpdate();
+    
                 return new Message(message_id, posted_by, message_text, time_posted_epoch);
             }
-            
+    
         } catch (SQLException e) {
             e.printStackTrace();
         }
-
+    
         return null;
     }
+    
 
     public Message updateMessageById(Message message) {
         Connection connection = ConnectionUtil.getConnection();
